pyproject.toml
[project]
name = "dreams-bot"
version = "0.1.0"
description = "Telegram bot for dream interpretation using LLM"
authors = [
    {name = "Dreams Bot Team"}
]
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
] 